coding

+++ backend/abstract_classes/abstract_specimen.hpp
@@ -19,7 +19,7 @@ public:
     ~AbstractSpecimen();
 
     virtual void evaluate() = 0;
-    virtual float mutate(AbstractSpecimen *specimen, uchar min, uchar max) = 0; 
+    virtual uchar mutate(AbstractSpecimen *specimen, uchar min, uchar max) = 0; 
     bool operator<=(AbstractSpecimen const &specimen2) const;
     mark_t getMark() const;
